requirements.txt
numpy==1.26.4
Pillow==11.0.0
torch==2.7.0+cu118
torchvision==0.22.0+cu118
tqdm==4.67.1
